Changed and improve few things
Changed square size and grid size, which is now fully dymanic.
Changed the max INT of the getRandomSequence function, it's now 60% of half the grid height.

Added a function setTerrainType which is recursive, it sets the grass and dirt given the number of the sequence. For instance, if the sequence number is 5, it'll set the terrain type starting from 5 square above or under the middle line and the recall itself with 4, 3, 2, etc. until this number is == 0.  I stopped at 0 because I have exceptions and conditions to have a nice terrain.

Thanks to this recursive function, I save a significant amount of lines and the bonus is that my terrain is fully dynamic. You just have to chnage the grid_width and grid_height variables.